move AvailableDatatypes to subfolder
