+++ client/src/components/AvailableDatatypes/AvailableDatatypes.vue
